chore: renamed router to index.js
