+++ dpp-frontend/src/router/index.js
@@ -21,14 +21,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import ErrorComponent from './components/general/ErrorComponent.vue';
+import ErrorComponent from "@/components/general/ErrorComponent.vue";
 import { createRouter, createWebHistory } from "vue-router";
-import HomeView from "./views/HomeView.vue";
-
+import HomeView from "@/views/HomeView.vue";
 
 const routes = [
     {
-        path: '/',
+        path: "/",
         name: "Home Page",
         component: HomeView,
     },
@@ -45,7 +44,7 @@ const routes = [
     {
         path: "/:id",
         name: "PassportView",
-        component: () => import("./views/PassportView.vue"),
+        component: () => import("@/views/PassportView.vue"),
     },
 ];
